[None][chore] consolidate bench tests into test_bench.py
Combine test_bench_data.py and test_bench_statistics.py with the new INT8
quant-choice test into a single tests/unittest/others/test_bench.py, per
review on #16245. All upstream tests are preserved verbatim.
